pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "CreditCardFraudDetection"
version = "0.0.1"
dependencies = [
    "pandas==2.2.2",
    "scikit-learn==1.5.1",
    "google-cloud-storage==2.18.2",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "tox==4.18.1",
    "black==24.8.0",
    "flake8==7.1.1",
    "mypy==1.11.2",
    "ipykernel==6.29.5",
]

